favour extra_implementations outside of the "extras" module, allowing them to override the extras module version
