+++ extras/fileformats/extras/application/serialization.py
@@ -1,3 +1,4 @@
+import json
 import tempfile
 import typing as ty
 from pathlib import Path
@@ -7,7 +8,13 @@ from pydra.compose import python
 
 from fileformats.application import Json, TextSerialization, Yaml
 from fileformats.application.serialization import SerializationType
-from fileformats.core import FileSet, converter, extra_implementation
+from fileformats.core import (
+    FileSet,
+    SampleFileGenerator,
+    converter,
+    extra_implementation,
+)
+from fileformats.core.exceptions import FormatMismatchError
 
 
 @converter(target_format=Json, output_format=Json)  # type: ignore[untyped-decorator]
@@ -42,3 +49,59 @@ def yaml_save(
 ) -> None:
     with open(yml.fspath, "w") as f:
         yaml.dump(data, f, **kwargs)
+
+
+@extra_implementation(FileSet.generate_sample_data)
+def generate_json_sample_data(
+    js: Json,
+    generator: SampleFileGenerator,
+) -> ty.List[Path]:
+    js_file = generator.generate_fspath(file_type=Json)
+    with open(js_file, "w") as f:
+        json.dump(
+            {
+                "a": True,
+                "b": "two",
+                "c": 3,
+                "d": [generator.rng.randint(0, 10), generator.rng.random(), 6],
+            },
+            f,
+        )
+    return [js_file]
+
+
+@extra_implementation(FileSet.generate_sample_data)
+def generate_yaml_sample_data(
+    yml: Yaml,
+    generator: SampleFileGenerator,
+) -> ty.List[Path]:
+    yml_file = generator.generate_fspath(file_type=Yaml)
+    with open(yml_file, "w") as f:
+        f.write(
+            f"""# Generated sample YAML file by FileFormats
+a: True
+b: two
+c: 3
+d:
+- {generator.rng.randint(0, 10)}
+- {generator.rng.random()}
+- 6
+"""
+        )
+    return [yml_file]
+
+
+@extra_implementation(FileSet.load)
+def load(jsn: Json, **kwargs: ty.Any) -> SerializationType:
+    try:
+        with jsn.open() as f:
+            dct: ty.Dict[str, ty.Any] = json.load(f, **kwargs)
+    except json.JSONDecodeError as e:
+        raise FormatMismatchError(f"'{jsn.fspath}' is not a valid JSON file") from e
+    return dct
+
+
+@extra_implementation(FileSet.save)
+def save(jsn: Json, data: SerializationType, **kwargs: ty.Any) -> None:
+    with jsn.open("w") as f:
+        json.dump(data, f, **kwargs)
