+++ fileformats/core/exceptions.py
@@ -39,3 +39,13 @@ class FileFormatsExtrasPkgUninstalledError(FileFormatsExtrasError):
 class FileFormatsExtrasPkgNotCheckedError(FileFormatsExtrasError):
     """If there is an "extra" hook in the datatype class and a extras package on PyPI
     but it hasn't been installed"""
+
+
+class FileFormatsClashingExtrasImplementationError(FileFormatsExtrasError):
+    """If there is an "extra" hook in the datatype class but no methods have been
+    registered on it"""
+
+
+class FileFormatsExtrasNotImplementedError(FileFormatsExtrasError, NotImplementedError):
+    """If there is an "extra" hook in the datatype class but no methods have been
+    registered on it"""
